requirements.txt
numpy
opencv-python-headless
Pillow
scikit-image
torch
torchvision
